pyproject.toml
[project]
name = "tp-microservice"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "psycopg2-binary",
    "pydantic",
    "pyzmq",
    "httpx",
]
